requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
catboost
xgboost
shap
joblib
tabulate
scipy
umap-learn
openpyxl